Returning function back to class
