+++ twitter_search/twitter_filtering/users_filtering/filter_users.py
@@ -99,6 +99,30 @@ class UserFilter:
             print("No previously classified users")
             self.unclassified_users = self.total_user_dict.copy()
 
+    @staticmethod        
+    def create_token(user):
+        """
+        This function will create the token to
+        classify the user using the HF model
+        """
+
+        token = " ".join(
+            [
+                (
+                    user["description"]
+                    if user["description"] is not None
+                    else ""
+                ),
+                (
+                    " ".join(user["tweets"])
+                    if user["tweets"] is not None
+                    else ""
+                ),
+            ]
+        )
+
+        return token
+
     def add_token_field(self, user):
         """
         Uses the 'create_token' function to add a new field
